+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+- Add `mysql_database` resource for creating, querying and removing databases
+- Add `mysql_user` resource for creating, modifying and removing database users
+- Remove Ubuntu 16.04 support
+
 ## 9.0.0 - *2020-11-27*
 
 - resolved cookstyle error: libraries/helpers.rb:201:9 convention: `Style/CommentAnnotation`